requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.1
django-auto-logout==0.5.1
django-axes==6.3.0
django-bootstrap-modal-forms==3.0.4
django-bootstrap3==23.4
django-crispy-forms==2.1
django-csp==3.7
django-dbbackup==4.0.2
django-easysettings==2.0.1
django-forms-bootstrap==3.1.0
django-password-policies==0.4.2
django-recaptcha==4.0.0
django-session-timeout==0.1.0
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
fontawesome-free==5.15.4
fontawesomefree==6.4.2
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
PyPDF2==3.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
